refactor(components): extract copy-to-clipboard toast into useCopyToast
QuickStartSection and MarkdownRenderer had byte-identical clipboard-copy
logic: a navigator.clipboard path, a textarea execCommand fallback, and an
auto-hiding confirmation toast.

Move it into a shared useCopyToast hook returning { toastVisible, handleCopy }.
